Move set_default_port
